Add an operator blocklist
The `tenzir.disabled-plugins` option is a list of names of plugins and
builtins to explicitly forbid from being used in Tenzir.

For example, adding `export` will prohibit use of the `export` operator
builtin, and adding `kafka` will prohibit use of the `kafka` connector
plugin. This allows for a more fine-granular control than the
`tenzir.allow-unsafe-pipelines` option.