Align store paths with store builder paths
Before this commit we relied on the filesystem implementation to
turn relative paths absolute, which created matching paths for the
store files at a very low level. This broke in the unit tests because
the in memory filesystem doesn't apply the same transformation.

Instead of reverting the builder to a relative path we now also
pass an absolute path to the passive store, because that moves us
ever so slightly closer to the desired state where the catalog manages
those locations directly.