Implement the `slice` operator
The `slice [<begin>]:[<end>]` operator cuts events in an interval $`[begin,
end)`$. A negative index counts from the end rather than from the
beginning, which is a syntax familiar to users coming with experience in
Python, jq, and many other languages.

Either `begin` or `end` can be omitted, and default to the 0 and the
size of the input, respectively.

`head N` is equivalent to `slice :N`.

`tail N` is equivalent to `slice -N:`.