Use a Docker volume for TENZIR_CACHE_DIRECTORY
This cache directory is the place where we create virtual environments
for the python operator, and this can take quite a bit of time. A shared
persistent volume for those environments eliminates this operator
startup delay for all but the first invokation of `docker run` with
the same python operator requirements parameter.