Allow setting the header for csv, tsv, and ssv manually
This adds a new `--header <header>` option to the `xsv` parser, and by
extension the `csv`, `tsv`, and `ssv` parsers. When set, the new option
effectively adds a line at the top of the input. This is useful when
parsing values from a file without a header, which is quite the common
occurrence.