Switch from JSON to MsgPack data transport
This commit changes how we receive data from Fluent Bit. Previously, we told
Fluent Bit that we want to receive data in JSON in the library callback.
Unfortunately this format isn't robust for large messages. By "robust", we mean
that we could get invalid JSON due to the following Fluent Bit code in
out_lib.c:184:

            len = snprintf(out_buf, out_size, "[%f,%s]",
                           flb_time_to_double(&tm),
                           buf);

Here, `buf` holds the Fluent Bit message and can contain cropped JSON that then
gets wrapped into the common array-style framing.

To avoid running into this problem we now go through a different code path in
Fluent Bit, avoiding JSON conversion alltogether. By switching the format of the
out_lib plugin to MsgPack, we now have direct access to the internal
representation of Fluent Bit that does not get cropped.

A beneficial side effect of this change is that we should have now higher
throughput as well due to the absence of an extra JSON roundtrip.