Further improve pipeline manager error handling
This fixes a case where the pipeline manager rendered unwanted
diagnostics, or in case of a back-and-forth translation between errors
and diagnostics invalid diagnostics.

Additionally, this fixes a race condition in the delivery of error-level
diagnostics from execution node, ensuring that diagnostics arrive at
most once at the analytics collector.