Fix a rare crash in the index on startup
This fixes a very rare crash in the index on startup where taxonomies
were not yet retrieved from the catalog but already attempted to access.

This crash was introduced a few months back when we backported a bug fix
for our CAF fork that allowed system messages to be scheduled while
another requests's response was awaited. The index uses the (rightfully)
deprecated CAF streaming, which uses system messages under the hood.
This means that awaiting the taxonomies on startup of the index did no
longer ensure that they couldn't be used from CAF streaming handlers
already.

Luckily, all this was rather easy to fix as we made taxonomies available
globally rather than just in the catalog actor quite a while ago. It
looks like we just forgot to update the logic of the index here.