Run functional tests when building the static binary
The Nix package expression is refactored into 2 derivations. The first
builds the binary and package as before, and the newly introduced second
derivation runs the functional tests. This allows us to iterate on the
tests without having to rebuild c++ code with every change.

The untested binary can still be accessed with the the `unchecked`
attribute, e.g. `...#tenzir-static.unchecked`, `...#tenzir-de.unchecked`
and so on.