Clean up the operator control plane
This is a first step towards making the control plane usable from other
threads, which is required to resolve tenzir/issues#1071.

Since this is a refactoring with a high merge conflict potential I opted
to put this into a PR of its own so that we can merge it quicker.