Fix hang for no-op transformations and sinks
If a transformation or sink operator finishes without yielding,
preceding operators effectively dangle because they are set up but never
receive any demand. We need to explicitly shut them down to avoid a
hang.

Note that there is no current operator that can run into this, so while
this is a bug fix it is not something that a user could've run into; I
just happened to run into this while working on a new operator with a
stub implementation that didn't yield anything.