Allow plugins to bundle further builtins
This change allows plugins to define further builtins that are
implicitly enabled as part of a plugin, allowing for a better
modularization of the code base.

To demonstrate that this change works, I've moved the two builtin
operators `enrich` and lookup` that were just aliases of the internal
operator `context enrich` and `context lookup`, respectively, into the
`context` plugin.