Fix BATS tests on macOS
This fixes an oversight in the porting of integration tests to BATS,
that was most likely to hit macOS users. In the test reference, there
was a directory that was commited twice—once with a titlecased and once
with lowercased test name.

To avoid this from happening in the future, I have changed the `check`
function in the test library to always use lowercase names for the test
reference directories.

Additionally, this fixes a set of tests that were marked as `skip` on
macOS only. Most of these happened simply because `zcat` isn't portable,
so I replaced them with the more widely available `gunzip -c`.

I also snuck in a fix for the YARA tests: this removes a lingering
compiled rule file in the source tree after running the tests.