Replace `tenzir.db-directory` with `tenzir.state-directory`
The option `tenzir.db-directory` is quite old, and comes from a time
when Tenzir wasn't all about pipelines, but rather mostly about its
storage engine VAST. Nowadays, we store lots of state in this directory
that is not part of a "database" in the classical sense. As such, the
option is now called `tenzir.state-directory`, with the old option name
still being supported for the foreseeable future in a
backwards-comaptible way.

Additionally, this change makes it possible to set the
`tenzir.cache-directory` on the command line when starting the node.

Fixes tenzir/issues#1109