Randomize pipeline ids
Why? Because of two very simple reasons:
1. Our integration tests should be order-independent. Having truly
   random pipeline ids forces us to write them without relying on the
   ids being incremental.
2. When using the `metrics` and `diagnostics` operators, it was rather
   easy to make the mistake `where pipeline_id == 42` when it should've
   been `where pipeline_id == "42"`. This won't happen anymore now that
   the id is obviously not a number anymore.

This change should not be user-visible, as we never made any guarantees
about the format of the pipeline ids.

Additionally, this fixes a small oversight in the recent addition of the
`created_at` and `last_modified` fields in managed pipelines, which were
not rendered as `time` values when using `show pipelines`.