Make FlatBuffers easier to use
This set of changes improves the developer experience when working with
FlatBuffers tables in Tenzir. I've made them out of frustration when
working with FlatBuffers for another plugin, and for an improved
reviewability I split them out into a separate pull request.

This PR
1. Fixes the version comparison to detect FlatBuffers versions without
   an opt-in to 64-bit offsets,
2. adds a convenience parameter for compiling, linking, and including
   FlatBuffers table in plugins,
3. makes globbing for files in plugins easier,
4. adds support for setting and checking the file identifier of a
   FlatBuffers root table, and
5. introduces support for size-prefixed FlatBuffers root tables.

None of this is user-facing, so no changelog entry required. The plugin
scaffolding changes are also entirely backwards-compatible.