Allow using `and`, `or`, and `not` in expressions
This change makes it so that `X or not (Y and Z)` is equivalent to `X ||
!(Y && Z)`. Since we want to move to that anyways with TQL2 and I'm
constantly mistyping this already, I figured it'd make sense to
introduce this four line diff that's entirely backwards-compatible to
allow using the new tokens for conjunctions, disjunctions, and negations
already.

This also removes the undocumented and untested `[+`, `[-`, `+]`, and
`-]` relational operators that were aliases for `in`, `not in`, `ni`,
and `not ni`, respectively. It's odd when you find code after over four
years of contributing to a project that's just silently been there and
that you've never seen. I doubt anyone is going to miss them.