Prevent tcp socket inheritance to child processes
Child processes created via `shell` or `python` inherit all currently
open file descriptors from the node, which prevent the tcp connector
form getting the same port upon restart, because it was now blocked by
the children, because of the inherited open file descriptors.

To prevent this, we now create our own socket with `FN_CLOEXEC` that we
assign to the asio abstraction where possible, and use `fcntl` where
not.