Make `remote python` work
This fixes `remote python`, i.e., using explicit location overrides with
the `python` operators to run a script at a node.

Additionally, this adds a 60 second timeout to waiting for the named
semaphore guarding the virtualenv setup. During development of this
branch, I managed to crash once while inside the critical section, and
the operator was unusable afterwards as it locked up indefinitely. It's
far from a perfect solution, but erroring after a minute is a lot better
than being stuck with no way out.

Fixes tenzir/issues#1632