Fix partial specialization of S3 configuration in URL
This fixes partial overrides of the S3 configuration in the arguments to
the URL. Previously, when providing any arguments to the S3 connector as
part of the URL, it failed to respect the default credentials from the
configuration file.

Additionally, this makes it possible to specify the S3 secrets in the S3
plugin's configuration file directly.