Introduce `--replace` and `--separate` for contexts
This adds `--replace` and `--separate` options to the `lookup` and
`enrich` operators.

The `--replace` option makes the operators replace the input value with
the context as opposed to adding it to the output.

The `--separate` option causes the operators duplicate the input events
for each matched field.

Additionally, this changes the context structure slightly to add the
meta information per context as opposed to once per event. This makes it
easier to work with the results of the two operators in downstream
tooling.