Revert recent changes to the `export` operator
This reverts the recent change to the `export` operatot that made it go
through the index actor, similar to the old `vast export` command. In
some production setups, we've seen the index be so busy sorting the
priority queue of partitions that something as simple as `export | head
1` takes minutes to finish.

We made these changes in the first place to reduce the pressure from
having too many queries running from _Threat Bus_ in parallel, but
that's alongside the matcher plugin on its way out and effectively
superseded by the `lookup` operator, so it should be safe to make this
change.