Fail late in the python operator setup
The python operator creates dedicated virtual envs on first use, and
that causes the initialization to take a long time. This causes problems
when the pipeline is created somewhere that has a short timeout,
triggering that before a response can be sent.

This change solves that by delaying the venv creation until the pipeline
has reached the running state, at which point client timeouts are
handled by the `/serve` endoint.