Support zero-copy reading in new Feather parser
This change builds on top of @balavinaithirthan's upcoming Feather
format and adds zero-copy support to it. This is a vast improvement when
reading Feather from a memory-mapped file, where copies of data can now
be avoided entirely until we need to do IPC or exchange any columns.

Added comments to code

Modify comments again

style modification