Add support for void-to-void operators
These operators are effectively commands, i.e., they cannot actually be
combined with any other operator, except for operator modifiers.

The rationale for this is an upcoming change to `context save`, `context
update`, and `context create` that turns the three operators into sinks.
But since `context create` already is a source operator this means that
we now had to implement support for operators that are both a source and
a sink at the same time. Luckily this was rather easy to implement.

For testing, I modified the `version` soucce operator to return a
`generator<std::monostate>` and to print the event on the stderr
instead. The only code outside of the execution nodes that I needed to
modify was the `every` operator modifier that ran into an assertion for
void-to-void operators.