Add an operator for dumping schemas
This is pretty helpful to get started with writing schemas, as it allows
one to go from inferred events over to having a manually specified
schema without needing to write everything from scratch.

The operator name is still a TBD and this definitely needs documentation
_iff_ we want to keep this. This originated in a discussion in the
community Discord, where a new user suggested this workflow.