Disable the `web` plugin on macOS
The Homebrew package for its http parser dependency was disabled quite a
while ago because the library is unmaintained and no longer compiled
from source against our current compiler toolchain, so it's really hard
to build the `web` plugin on macOS. This just disables it.

This also re-enables the `parquet` plugin on macOS, as that no longer
uses illegal instruction as far as I can tell.