Add timeout options to `summarize`
This adds two new options to the `summarize` operator:
- `timeout <duration>` specifies the maximum lifetime of a bucket from
  the time it was created at.
- `update-timeout <duration>` specifies the maximum lifetime of a bucket
  from the time it was last updated.

These options basically enable streaming aggregations. The `timeout`
option is useful to guarantee that events are held back no more than the
specified duration. The `update-timeout` is useful when operating on
a stream of events where it's known that events that would be sorted
into the same group arrive at near the same time.