Remove stream managers when decommissioning partitions
We noticed that in some deployments, the CPU usage of the `tenzir.index`
thread would rise over time. Most time was spent in
`std::_Rb_tree_increment`—which is an internal of `std::map` or
`std::set`. But we didn't use that anywhere in the index actor, or so we
thought: it turned out that CAF itself uses a `std::map` for its stream
managers, creating a new one for every active partition. When
decommissioning an active partition into a passive partition, that entry
in the map was not removed, making further interactions with the map
increasingly expensive.

This change makes it so we manually erase stale stream managers when
decommissioning an active partition. This really should be considered a
bug in CAF, but since upstream has replaced the streaming functionality
entirely it doesn't make much sense to report or fix it there.