Do not persist pipeline state updates on node shutdown
This fixes a bug in pipeline management that caused the states of
pipeline to be persisted even when the node was already shutting down.

Shutting down a node naturally stops some source operators. Pipelines
with these sources naturally complete when the node shuts down. However,
from a user perspective, these pipelines should run again when the node
starts up again, as they were running when the node was stopped.

To solve this problem, the pipeline manager component no longer writes
state updates for pipelines _after_ the node shutdown sequence started.