Force error diagnostics to propagate immediately
This change makes it so that
`diagnostic::error(...).emit(ctrl.diagnostics())` throws an exception
containing the diagnostic as opposed to sending the diagnostic. This is
less error prone, because now the caller does not need to ensure that
they yield control to the execution node's runloop immediately after
emitting an error. The exception will just do that automatically.