Make submodules easier to work with
For submodules other than `contrib/tenzir-plugins`, this sets the
`shallow = true` option, which causes `git submodule update` to do a
shallow clone of submodules, unless `--no-recommend-shallow` is used.

For `contrib/tenzir-plugins`, this explicitly disables shallow cloning,
and sets the merge strategy to rebase in the submodule in case of
uncomitted changes. This makes rebasing across submodule bumps a lot
easier.