Create a variant of the tenzir-demo image for use in the platform CI
The existing tenzir-demo image has delays of 10s and more on weak
machines when processing pipeline launch and update endpoints, because
the pipeline manager competes with the running imports for resources.
This leads to sporadic failure of our integration tests.

Since rewriting the scheduler to fix this would be a huge undertaking,
we instead build a variant of the demo image with less things going on.