Tone down execution node backoff behavior
This turns down the execution node backoff behavior. Instead of going
from 10ms to 1s with 1.25x increases, we now go from 250ms to 1 min with
2x increases.

Additionally, this causes execution nodes that explicitly ask not to be
scheduled to be exempt from backoff and not to be scheduled at all.