Fix `offset{}.get(…)` for struct arrays and record batches
This fixes access into struct arrays and record batches for empty
offsets. This is handled nicely already for table slices, but it looks
like we forgot this overload. I ran into this while working on some
other code, and figured it'd be worth extracting this fix already. It
shouldn't be possible for users to run into this before TQL2.