Use custom checked math functions and fix type of multiplications
The previous implementation did not return `signed` for `signed - unsigned` or
multiplications that involve `signed`. Furthermore, it seems to have bug in the
definition of `subtract`. Since I did not find a library that does what we want,
I ultimately decided to implement this myself. Hopefully it doesn't contain too
many bugs. In any case, this can be improved later.