Simplify the node actor setup
This removes a bunch of old, no longer needed code relating to the node
actor, which is the central component of the `tenzir-node` process.

In particular, this removes:
- Dynamic spawning of components via the component factory, which was
  long superseded by component plugins; the last remaining components
  outside of component plugins are now spawned by the node actor itself,
  which makes them easier to remove in the future
- All but one way of accessing components through the component registry
- Large parts of the node actor interface that were unnecessary
- The `--node,-N` option for the deprecated `tenzir-ctl` binary

This brings us one step closer to removing the built-in database
component, and also brings us one step closer to removing `tenzir-ctl`.

Additionally, this gracefully removes a options that are no longer used,
and the `vast` symlink for the `tenzir-ctl` binary.