Ensure plugins are always in the right order
This fixes a rather obscure bug that I've been trying to hunt down for
quite some time, where some plugins are not stored sorted by name.
Turns out this was only happening for builtins and depended on the
order of static initialization. Plugins may bundle further builtins,
which then get added once the plugin is loaded.