Reduce load in `metrics --live` and `diagnostics --live`
Before this change, `export --internal --live` (which backs `metrics
--live` and `diagnostics --live`) received all imported events, not just
internal events. This change makes it so that this the internal
attribute is evaluated earlier.

To get a good feel of the reduction, use the following pipelines:

```
export --live
| summarize eps=count(.) timeout 1s
| summarize mean(eps) timeout 1m
```

```
export --live --internal
| summarize eps=count(.) timeout 1s
| summarize mean(eps) timeout 1m
```

Previously, the `export` operators needed to evaluate the sum of the
events coming in, now they don't anymore.