Raise the default and max timeouts for `/serve`
This brings up the default and maximum long-polling timeouts for
`/serve` from 2s and 5s to 5s and 10s, respectively.

I've made this change because we've started polling a lot of metrics and
diagnostics from the app, and this drastically reduces the load on the
API when there are no diagnostics.

We can't really raise it much higher because of AWS' lambda and
websocket gateway timeouts, but this I think we can safely do.