Break back pressure in hidden subscribers
This change makes it so that `subscribe` no longer propagates back
pressure when run as part of a hidden pipeline, i.e., a pipeline that is
not visible on the Tenzir Platform.

Without this change, it was surprisingly easy to break publishers by
simply subscribing to their topic and piping the data into an operator
that did nothing, e.g., `serve` with a very small buffer size.