Add the `buffer` operator for breaking back pressure
The `buffer` operator buffers up to the specified number of events in an
in-memory buffer.

By default, operators in a pipeline run only when their downstream
operators want to receive input. This mechanism is called back pressure.
The `buffer` operator effectively breaks back pressure by storing up to
the specified number of events in memory, always requesting more input,
which allows upstream operators to run uninterruptedly even in case the
downstream operators of the buffer are unable to keep up. This allows
pipelines to handle data spikes more easily.

Since back pressure is _between_ operators, the `buffer` operator
expands to two internal operators under the hood:
`internal-write-buffer` and `internal-read-buffer`. The former writes
into a buffer and always asks for more input, and the latter reads from
the buffer and is considered input-independent because it receives
events from the shared buffer rather than from its counterpart.