Make the `serve` operator cheaper to start
When using Tenzir via [app.tenzir.com](https://app.tenzir.com), the
`serve` operator is one of the must used operators, as all data transfer
from pipelines to the Tenzir Platform runs through it.

The operator was—for the lack of better internal APIs at the time of
writing it—marked as detached, i.e., an operator that did not
participate in cooperative scheduling but rather required its own
thread.

This change makes it so the operator uses the newer internal APIs that
work well with cooperative scheduling. This has two effects: spawning a
pipeline with `serve` is now faster, and there's less oversubscription
of threads.

Additionally, this ports the `serve` operator to TQL2.