Implement the `cache` operator
`cache` is a transformation that passes through events, creating an
in-memory cache of events on the first use. On subsequent uses, the
operator signals upstream operators no to start at all, and returns
the cached events immediately.

Caches are identified through a unique id, which is an arbitrary string.

Caches have a capacity limit, emitting a warning when the limit is
reached. The capacity defaults to 4Mi events.

Caches have a TTL defining the maximum inactvitiy time until the cache
is evicted. The TTL starts when writing the cache completes (or runs
into the capacity limit), and resets whenever the cache is read from.
THe TTL defaults to 1 minute.

The optional `mode` parameter can be set to either `read`, `write`, or
`readwrite` to change the operator into a source, sink, or
transformation, respectively. The option defaults to `readwrite`. If the
option is set to `read` the cache must exist already, and if it is set
to `write` the cache must _not_ exist already.