Fix a crash when starting `export` on shutdown
We've seen a few reports of crashes on main where the `export` operator
fails to retrieve the `filesystem` component's actor handle. This cannot
happen on startup, as the filesystem is registered before the node has
started up, but it can very well happen on shutdown, if the node is
tearing down while a new `export` operator is starting up. This fixes
the crash by responding with a proper error message instead when a
component is not available.