Perform individual catalog lookups in `lookup`
This makes `lookup` faster by performing one catalog lookup per event in
the context update, as opposed to one catalog lookup per batch of
events. This may seem a bit counterintuitive, but the large disjunction
of expressions for batched updates caused a lot of false positives.
This new approach has less of them, but on the other hand may cause
duplicate results to be returned from the catalog. To offset this, the
`lookup` operator now merges expressions into a disjunction when it
notices repeat partitions being returned as candidates.

This also fixes an issue where a failed catalog lookup caused the
`retro.queued_events` metric to miscount until after the node restarted.