Make `tenzir` and `tenzir-node` start faster
We noticed that flamegraphs show a large amount of time on startup being
spent in the function `tenzir::config_dirs(…)`. Apparently, the function
was called $`N+1`$ times for $`N`$ plugins and builtins available.
That's a bit unnecessary, so I shuffled the code around to only call the
function exactly once.