Fix a logic error for retried requests in `/serve`
This fixes a logic error for retried requests in `/serve`. The endpoint
already gracefully handled repeated continuation tokens, but only after
the initial response for that continuation token was returned. In case
the request to `/serve` is cancelled and then immediately retried, we
returned an error if the second request for the same continuation token
arrived before the first one (that was cancelled) was replied to.

From the perspective of the Tenzir Platform, this fixes an issue where
hitting stop and then continue shortly after one another in the Explorer
would show an error.