Stabilize the `bitz` format
This changes the `bitz` format to no longer rely on CAF's type ids
internally, and instead be a stream of size-prefixed Arrow IPC messages.
This breaks the format once, but makes it less likely to break across
versions in the future.

Additionally, this fixes the zero-copy adapter from our `chunk_ptr` to
Arrow's equivalent `arrow::Buffer`. Using this sometimes caused the
`feather` and `parquet` parsers to fail with a segmentation fault deep
within Arrow's `arrow::util::ReferencedBufferSize` function because it
tried to get the size of the buffer that had already been deallocated.

Lastly, this adds `read_bitz` and `write_bitz` for TQL2.