Fix double-closing fds in the python operator
In boost::process, a pipe's file descriptors are closed in its
destructor.

When they were already closed, this results in a harmless "Bad file descriptor"
result in the best case when attempting to close an already-closed fd.

In the worst case, the descriptor was already re-opened in the meantime
and we close an unrelated file descriptor.