Raise max fd soft limit during initialization
We often run into issues around the low default of maximum open
file descriptors. On current Linux systems the soft limit is
1024, but processes can raise it up to a hard limit set by the
kernel or parent process if it knows that it is safe to do so.

For our own codebase we know that it is safe, so we can eliminate
this stumbling block.

Technically we should reset it to 1024 or lower after forking of
any subprocesses, but this is not easily doable with boost process,
so we refrain from that and consider non-supporting legacy software
as unsupported for the `shell` operator.