Add the `aggregation` context type
This implements a new `aggregation` context type. It functions much like
a `lookup-table` context, except that it takes as a required argument a
list of assignments of aggregation function calls.

When updating the context, the provided key field is used to identify
which set of aggregations to update based on the data. When enriching
with the context, the aggregation instances are used to create the
enrichment.

Here's an example. Consider a context like this:

```yaml
contexts:
  users:
    type: aggregation
    args:
      fields:
        - first_seen=min(timestamp)
        - last_seen=max(timestamp)
        - ips=distinct(ip_address)
```

You update the context from a stream of events like this:

```json
{
  "username": "dominiklohmann",
  "timestamp": "2024-10-15T16:07:00.843514",
  "ip": "1.2.3.4"
}
```

In another pipeline, you can now enrich usernames with this information,
automatically adding the `first_seen`, `last_seen`, and `ips` fields
defined in the context based on all previous context updates.

> [!NOTE]
> This also makes three somewhat related adjustments to surrounding code:
> - We now forbid creating a `session` from a temporary
>   `session_provider`, as that is inherently unsafe.
> - `context::make_query` no longer exists. It was effectively dead code
>   since the removal of `lookup --snapshot`.
> - Getting the result of an `aggregation_instance` may no longer modify
>   the instance to guarantee that it is safe to get an intermediate
>   result.