Return instantly in `/serve` if pipeline fails early
Before this change, the `/serve` endpoint waited for its long-poll
timeout to expire if the pipeline failed before the first request to
`/serve` came in. Now it returns instantly. For the Tenzir Platform,
this causes the loading button in the Explorer to disappear immediately
for such cases.

Here's a simle reproduction that now returns instantly but took some
time before:

```
// tql2
shell "echo null"
read_json
serve "foo"
```

```
// tql2
api "/serve", {
  serve_id: "foo"
}
```