Ensure cache directory is writeable for multiple users
The `tenzir` binary creates intermediate directories for the cache
directory as necessary. If these do not yet exist, they will be owned by
the first user running the process. That can cause other users running
the process later to fail, as they may not have permission to access the
intermediate directories if the system is locked down enough. This
change fixes this problem by simply removing the intermediate
directories.