Allow reinstantiating the `buffer` operator
This fixes a rather annoying bug: it was not possible to instantiate the
`buffer` operator a second time without parsing it again. This is due to
how the operator, which physically consistents of two operators,
exchanges information by (ab)using the actor registry as a communication
channel. This broke when (1) restarting a pipeline with the oeprator
through the Tenzir Platform, or (2) using the operator in a nested
pipeline that gets instantiated multiple times, for example in
`load_tcp` with at least two connections.