Specify the type of the `version` operator explicitly
We often use the `version` operator as a dummy event in testing, and
while specifying the type explicitly has no practical impact, it makes
it so that the schema id stays constant. We had just changed that from
`list<string>` to `list<null>` accidentally with the removal of old
feature flags, as the list's element type could no longer be inferred,
and that broke a test setup for the Tenzir Platform.