Gracefully handle duplicate serve requests
If a http request to `serve` get's re-sent with the same continuation
token it is possible that the second one does not fully dispose the
first when the clock already inserted the scheduled action in the
serve manager's mailbox.
That means asserting that the continuation token from the first request
must still be valid after the timeout is wrong.